Settings Page
Ive created 2 settings page. one for the user,just to change personal details and choose if he wants to be an admin. One for the admin who will also be able to change personal details and choose if he wantsto be a user.The admin settings page will allow also to edit category question and answer or to delete it.There are 3 comboBoxes to choose and text fields to put the new answer or question.If you want we can put another text field for the category.My problem is that I m trying to add the options for each drop down menu but it doesnt work.There are 2 new methods openSettingsAdmin and 1 for user.For now only the admin method is connected with the settings menu.