sxore-reports
So, Ive put the images inside the src file, I tried to put it outside and change the path in css but it didnt work,you can always try as well. I ve create a small window for the score but i want to put it in the same scene with the quiz without closing the quiz page. also if we want to add things to the report tell so i can change it.